@petermattis 2019-11-05
TFTR!

The `MVCCScan` benchmarks had an anomaly with `valueSize=64`. I'm not sure if that is the one you're recalling. The problem here appears to be different upon initial inspection.

bors r+